# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Boite \u00e0 id\u00e9e",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/projects": {
            "get": {
                "tags": [
                    "projects"
                ],
                "summary": "R\u00e9cup\u00e9rer tous les projets",
                "responses": {
                    "200": {
                        "description": "Tous les projets"
                    }
                }
            },
            "post": {
                "tags": [
                    "projects"
                ],
                "summary": "Ajouter un projet",
                "parameters": [
                    {
                        "name": "name",
                        "in": "body",
                        "description": "Le nom du projet",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "string"
                    },
                    {
                        "name": "image_link",
                        "in": "body",
                        "description": "Le lien de l'image du projet",
                        "required": true,
                        "schema": {
                            "properties": {
                                "image_link": {
                                    "type": "string",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "string"
                    },
                    {
                        "name": "id_creator",
                        "in": "body",
                        "description": "L'id de l'utilisateur qui ajoute le projet",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id_creator": {
                                    "type": "integer",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne le projet avec un id"
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "tags": [
                    "projects"
                ],
                "summary": "R\u00e9cup\u00e9rer un projet via son id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'id du projet",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "default": "1"
                                }
                            }
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne le projet correspondant \u00e0 l'id en param\u00e8tre"
                    }
                }
            },
            "put": {
                "tags": [
                    "projects"
                ],
                "summary": "Modifier un projet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'id du projet",
                        "required": true,
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "default": "1"
                                }
                            }
                        },
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "body",
                        "description": "Le nom du projet",
                        "required": false,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "string"
                    },
                    {
                        "name": "image_link",
                        "in": "body",
                        "description": "Le lien de l'image du projet",
                        "required": false,
                        "schema": {
                            "properties": {
                                "image_link": {
                                    "type": "string",
                                    "default": "Test"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "projects"
                ],
                "summary": "Supprimer un projet via son id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "L'id du projet \u00e0 supprimer",
                        "required": true,
                        "type": "integer",
                        "minimum": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projet supprim\u00e9 avec succ\u00e8s"
                    },
                    "404": {
                        "description": "Projet introuvable"
                    }
                }
            }
        },
...
    },
...
}